JPA repository, dependency injection, database seeding

+++ src/main/java/com/nilsonsantiago/course/entities/User.java
@@ -17,7 +17,7 @@ public class User implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	private String email;
 	private String phone;
@@ -29,7 +29,7 @@ public class User implements Serializable {
 		
 	}
 	
-	public User(Integer id, String name, String email, String phone, String password) {
+	public User(Long id, String name, String email, String phone, String password) {
 		this.id = id;
 		this.name = name;
 		this.email = email;
@@ -37,11 +37,11 @@ public class User implements Serializable {
 		this.password = password;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
